requirements.txt
discord.py
ffmpeg
yt-dlp
python-dotenv
pynacl